Time: 9 ms (70.53%), Space: 9.9 MB (78.95%) - LeetHub

+++ 260-single-number-iii/260-single-number-iii.cpp
@@ -3,7 +3,7 @@ public:
     vector<int> singleNumber(vector<int>& nums) {
          int n = nums.size();
         
-        long long a = nums[0];
+        long a = nums[0];
         
         for(int i=1;i<n;i++){
             a^=nums[i];
